src/requirements.txt
PyYAML
APScheduler
psutil
paho-mqtt
Adafruit-ADS1x15
Adafruit-SHT31
luma.oled
Pillow
SMBus
RPi.GPIO
simple-pid
requests
